+++ android/app/src/main/java/cn/toside/music/mobile/MainActivity.java
@@ -1,14 +1,11 @@
 package cn.toside.music.mobile;
 
+import com.facebook.react.ReactRootView;
 import com.reactnativenavigation.NavigationActivity;
 import com.facebook.react.ReactActivityDelegate;
-import com.facebook.react.ReactRootView;
+import com.facebook.react.defaults.DefaultReactActivityDelegate;
 
 public class MainActivity extends NavigationActivity {
-
-
-
-
 
   public static class MainActivityDelegate extends ReactActivityDelegate {
     public MainActivityDelegate(NavigationActivity activity, String mainComponentName) {
@@ -30,4 +27,21 @@ public class MainActivity extends NavigationActivity {
       return BuildConfig.IS_NEW_ARCHITECTURE_ENABLED;
     }
   }
+
+  /**
+   * Returns the instance of the {@link ReactActivityDelegate}. Here we use a util class {@link
+   * DefaultReactActivityDelegate} which allows you to easily enable Fabric and Concurrent React
+   * (aka React 18) with two boolean flags.
+   */
+  // @Override
+  // protected ReactActivityDelegate createReactActivityDelegate() {
+  //   return new DefaultReactActivityDelegate(
+  //       this,
+  //       getMainComponentName(),
+  //       // If you opted-in for the New Architecture, we enable the Fabric Renderer.
+  //       DefaultNewArchitectureEntryPoint.getFabricEnabled(), // fabricEnabled
+  //       // If you opted-in for the New Architecture, we enable Concurrent React (i.e. React 18).
+  //       DefaultNewArchitectureEntryPoint.getConcurrentReactEnabled() // concurrentRootEnabled
+  //       );
+  // }
 }