+++ src/core/init/i18n.ts
@@ -1,6 +1,6 @@
 import { createI18n } from '@/lang/i18n'
 import type { I18n } from '@/lang/i18n'
-import { deviceLanguage } from '@/utils/tools'
+import { getDeviceLanguage } from '@/utils/tools'
 import { setLanguage, updateSetting } from '@/core/common'
 
 
@@ -10,6 +10,7 @@ export default async(setting: LX.AppSetting) => {
   global.i18n = createI18n()
 
   if (!lang || !global.i18n.availableLocales.includes(lang)) {
+    const deviceLanguage = await getDeviceLanguage()
     if (typeof deviceLanguage == 'string' && global.i18n.availableLocales.includes(deviceLanguage as I18n['locale'])) {
       lang = deviceLanguage as I18n['locale']
     } else {