+++ src/utils/nativeModules/utils.ts
@@ -41,3 +41,6 @@ export const writeFile = async(filePath: string, data: string): Promise<void> =>
 export const readFile = async(filePath: string): Promise<string> => {
   return UtilsModule.getStringFromFile(filePath)
 }
+export const getSystemLocales = () => {
+  return UtilsModule.getSystemLocales()
+}